Add `Delta (Maybe …)` instance and `modifyDBMaybe`
… to cover the case where we want to make no modifications.

Even though `Maybe …` is now an instance of `Delta`, it is not useful to use this as the base delta type for a DBVar, hence the new helper function.

That said however, one could try to add a function that lifts `DBVar m da` into `DBVar m (Maybe da)` by introducing the concept of "delta transformer". `Maybe` would be one example of this. 🤔